Rename bundles and controllers for exercises 2 and 3
