+++ ex03/config/bundles.php
@@ -1,7 +1,7 @@
 <?php
 
 return [
-    App\Ex03Bundle\Ex03Bundle::class => ['all' => true],
+    App\E03Bundle\E03Bundle::class => ['all' => true],
     Symfony\Bundle\FrameworkBundle\FrameworkBundle::class => ['all' => true],
     Doctrine\Bundle\DoctrineBundle\DoctrineBundle::class => ['all' => true],
     Doctrine\Bundle\FixturesBundle\DoctrineFixturesBundle::class => ['dev' => true, 'test' => true],
